Export de la lista de items a un componente a parte y creacion del componente dinero
