+++ components/Money.js
@@ -1,0 +1,12 @@
+import MoneyIcon from '@material-ui/icons/Money';
+import Paper from '@material-ui/core/Paper';
+
+const Money = () => (
+    <Paper>
+        <div>
+        <p><MoneyIcon/> Cantidad de dinero </p>
+        </div>
+    </Paper>
+);
+
+export default Money;
